# docs/api-v1-openapi.yaml
openapi: 3.1.0
info:
  title: CloudOps API v1
  version: 1.0.0
  description: >-
    Local Owner commands and bounded MySQL-backed CloudOps projections.
    Numeric database identifiers, secrets, leases, checkpoints, and raw prompts
    are never public.
servers:
  - url: /
paths:
...
  /api/v1/agent/consultations:
    get:
      operationId: listAgentConsultations
      description: Durable Consultation history without hidden cross-session chat injection.
      parameters: [{ $ref: "#/components/parameters/Limit" }]
      responses:
        "200": { description: Consultation summaries and active snapshot identities }
        "501": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      operationId: createTelemetryConsultation
      description: Local Owner creates a Consultation with one immutable Context Snapshot from exact query and Evidence references.
      parameters: [{ $ref: "#/components/parameters/Origin" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "201": { description: Consultation and immutable initial Context Snapshot }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/agent/investigations:
    get:
      operationId: listAgentInvestigations
      description: Durable Alert and Incident Investigation history.
      parameters: [{ $ref: "#/components/parameters/Limit" }]
      responses:
        "200":
          description: Investigation summaries
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/AgentRun" } }
        "501": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/agent/investigations/{id}:
    parameters: [{ $ref: "#/components/parameters/PublicID" }]
    get:
      operationId: getAgentInvestigation
      description: Exact Investigation with immutable context, bounded tool progress, Evidence citations, and outcome.
      responses:
        "200": { description: Investigation detail, content: { application/json: { schema: { $ref: "#/components/schemas/AgentRun" } } } }
        "404": { $ref: "#/components/responses/Problem" }
        "501": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/agent/investigations/{id}/cancel:
    parameters: [{ $ref: "#/components/parameters/PublicID" }]
    post:
      operationId: cancelAgentInvestigation
      description: Local Owner requests bounded cancellation without deleting persisted progress.
      parameters: [{ $ref: "#/components/parameters/Origin" }]
      responses:
        "202": { description: Cancellation requested or completed }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/agent/consultations/{id}:
    parameters: [{ $ref: "#/components/parameters/PublicID" }]
    get:
      operationId: getAgentConsultation
      description: Durable messages and explicit immutable Context Snapshots for one Consultation.
      responses:
        "200": { description: Consultation detail }
        "404": { $ref: "#/components/responses/Problem" }
        "501": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
...
  /api/v1/knowledge-items:
    get:
      operationId: listKnowledgeItems
      description: Owner-confirmed, versioned Knowledge Items with scope and review state.
      parameters: [{ $ref: "#/components/parameters/Limit" }]
      responses:
        "200": { description: Knowledge Item page }
        "501": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      operationId: createKnowledgeItem
      description: Local Owner confirms exact content before it becomes cross-session guidance.
      parameters: [{ $ref: "#/components/parameters/Origin" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        "201": { description: Owner-confirmed Knowledge Item revision }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/knowledge-items/{id}:
    parameters: [{ $ref: "#/components/parameters/PublicID" }]
    get:
      operationId: getKnowledgeItem
      description: Exact Knowledge Item and immutable revision history.
      responses:
        "200": { description: Knowledge Item detail }
        "404": { $ref: "#/components/responses/Problem" }
        "501": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    patch:
      operationId: reviseKnowledgeItem
      description: Local Owner creates a new immutable revision or changes lifecycle state.
      parameters: [{ $ref: "#/components/parameters/Origin" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        "200": { description: Revised Knowledge Item }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteKnowledgeItem
      description: Local Owner marks the item deleted while retained exact revisions remain explainable.
      parameters: [{ $ref: "#/components/parameters/Origin" }]
      responses:
        "204": { description: Knowledge Item marked deleted }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/runbook-guidance:
    get:
      operationId: listRunbookGuidance
      description: Git-managed Runbook Guidance with exact content revision; never current Evidence or write authority.
      responses:
        "200": { description: Versioned Runbook Guidance page }
        "501": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  parameters:
    PublicID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid, minLength: 36, maxLength: 36 }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
    Origin:
      name: Origin
      in: header
      required: true
      schema: { type: string, format: uri, maxLength: 2048 }
...
  schemas:
...
    AgentRun:
      type: object
      description: Durable Agent run projection. Scenario identity is inherited from the immutable Context Snapshot.
      required: [id, subject_type, configuration_revision_id, context_snapshot_id, status, uncertainty, objective, prompt_version, created_at, updated_at, evidence_count, steps, evidence_citations, guidance_citations, action_cards, operation_plans]
      properties:
        id: { type: string, format: uuid }
        subject_type: { type: string, enum: [alert, incident, consultation] }
        alert_id: { type: string, format: uuid }
        incident_id: { type: string, format: uuid }
        consultation_id: { type: string, format: uuid }
        configuration_revision_id: { type: string, format: uuid }
        context_snapshot_id: { type: string, format: uuid }
        scenario_id: { type: string, pattern: "^scenario-[a-z0-9]([-a-z0-9]*[a-z0-9])?$", maxLength: 63 }
        status: { type: string, enum: [pending, running, completed, failed, cancelled] }
        outcome: { type: string, enum: [diagnosed, insufficient, cancelled, failed] }
        uncertainty: { type: string }
        objective: { type: string }
        answer: { type: string }
        model_provider: { type: string }
        actual_model: { type: string }
        prompt_version: { type: string }
        tool_schema_version: { type: string }
        failure_code: { type: string }
        failure_summary: { type: string }
        cancel_requested_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        evidence_count: { type: integer, minimum: 0 }
        steps: { type: array, items: { type: object } }
        evidence_citations: { type: array, items: { type: object } }
        guidance_citations: { type: array, items: { type: object } }
        action_cards: { type: array, items: { type: object } }
        operation_plans: { type: array, items: { $ref: "#/components/schemas/OperationPlan" } }
    OperationTarget:
      type: object
      additionalProperties: false
      required: [cluster_id, environment, namespace, workload_kind, workload_name]
      properties:
        cluster_id: { type: string, minLength: 1, maxLength: 128 }
        environment: { type: string, minLength: 1, maxLength: 64 }
        namespace: { type: string, minLength: 1, maxLength: 63 }
        workload_kind: { type: string, enum: [Deployment] }
        workload_name: { type: string, minLength: 1, maxLength: 253 }
        scenario_id: { type: string, pattern: "^scenario-[a-z0-9]([-a-z0-9]*[a-z0-9])?$", maxLength: 63 }
...
    ActionAuthorization:
      type: object
      additionalProperties: false
      required: [id, subject_type, subject_id, authorized_content_hash, authorized_by, reason, expires_at, created_at]
      properties:
        id: { type: string, format: uuid }
        subject_type: { type: string, enum: [action_card, operation_plan] }
        subject_id: { type: string, format: uuid }
        authorized_content_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        authorized_by: { type: string }
        reason: { type: string }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    OperationPlan:
      type: object
      additionalProperties: false
      required: [id, run_id, configuration_revision_id, authority, operation_type, target, parameters, intended_state, preconditions, risk, verification_intent, content_hash, status, expires_at, created_at]
      properties:
        id: { type: string, format: uuid }
        run_id: { type: string, format: uuid }
        configuration_revision_id: { type: string, format: uuid }
        authority: { type: string, const: high_impact }
        operation_type: { type: string }
        target: { $ref: "#/components/schemas/OperationTarget" }
        parameters: { type: object }
        intended_state: { type: object }
        preconditions: { type: array, items: { type: object } }
        risk: { type: string }
        verification_intent: { type: object }
        content_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        status: { type: string, enum: [proposed, authorized, expired, cancelled] }
        expires_at: { type: string, format: date-time }
        authorization: { $ref: "#/components/schemas/ActionAuthorization" }
        created_at: { type: string, format: date-time }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, request_id, trace_id, next_steps]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        request_id: { type: string }
        trace_id: { type: string }
        next_steps:
          type: array
          minItems: 1
          items: { type: string }
  responses:
...
    Problem:
      description: Stable problem response
      headers:
        X-Request-ID: { schema: { type: string } }
        X-Trace-ID: { schema: { type: string } }
        Idempotent-Replay: { schema: { type: string, enum: ["true"] } }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }